Fix dimensionality mismatch in z and t within tracksfromOPMD
tracksfromOPMD has a feature which removes iterations of its variables which contain invalid values. However, it does not remove the corresponding iterations of t. Thus the line
"f[f'tracks/{i_tr:d}/z'] = z + c * t" can cause a dimensionality error.

This commit adds an output "good_inds" to the "split_tracks_by_nans" function, which allows us to truncate the list of t, according to which t values correspond to valid variables. This is then implemented in tracksfromOPMD to avoid the dimensionality error.